Update README.md
#  Battle Stocks README


# Project Overview
In Battle Stocks, the user is given a list of stocks fetched from IEX API as well as the company information and website to do more research. The user then bets on whether the share price increased/decreased over the past year, what today's price is and how much they are willing to bet. If the user is half right or completely right, the user can then purchase the stock at it s current price and it becomes part of their portfolio. Prices are always updating as fetched data is constantly taken from IEX.